refactor: move visualization into new directory, split matplotlib functionality
