+++ tests/visualize/reference/ranking.pdf
