+++ tests/visualize/reference/scan.pdf
